synthetic_data_genration/requirements.txt
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
tqdm>=4.66.0
google-auth>=2.23.0
